Use ID for LabProcedure status instead of enum in DB
Via discussion with xmiao2, we concluded that defining status as an enum
in MySQL is bad practice. This replaces that construct with an ID in the
DB table definition, and gives LabProcedure an enum that is settable by
both ID and name. This eliminates the need for a 3rd table to keep track
of valid statuses.